Added functionality for changeUtil module, and completed user interaction with error handling

+++ changeUtil.h
@@ -1,2 +1,2 @@
 
-void calcChange(float dollarsOwed, float dollarsPaid, int change[]);
+void calcChange(double dollarsOwed, double dollarsPaid, int change[]);
